requirements.txt
streamlit
pandas
google-api-python-client
google-auth
beautifulsoup4
requests
googlesearch-python
